requirements.txt
fastapi
uvicorn[standard]
pymongo
requests
pydantic-settings
python-dotenv
chromadb